API/requirements.txt
flask
mysql-connector-python==8.0.25
